requirements/server_docker.txt
-r common.txt

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5

# Database
mysqlclient==2.2.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Federated Learning
flwr>=1.11.0
grpcio>=1.59.0
protobuf>=4.25.0

# Machine Learning
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.3.2

# ML Experiment Tracking
mlflow==2.8.1

# Storage
boto3==1.29.7
django-storages==1.14.2

# Privacy
opacus==1.4.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
openpyxl==3.1.5

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Utilities
python-multipart==0.0.6
python-json-logger==2.0.7
psutil==5.9.6

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# API Documentation
drf-spectacular==0.26.5
